[Gold IV] Title: 파일 합치기 3, Time: 2736 ms, Memory: 281032 KB -BaekjoonHub

+++ 백준/Gold/13975. 파일 합치기 3/파일 합치기 3.java
@@ -14,7 +14,7 @@ public class Main {
 
             st = new StringTokenizer(br.readLine());
             for (int j = 0; j < k; j++) {
-                pq.add((long) Integer.parseInt(st.nextToken()));        // 업캐스팅
+                pq.add(Long.parseLong(st.nextToken()));        
             }
 
             long sum = 0;
